Revert "chore: move cursor bubble to internal"
This reverts commit a4ed54327f08e510ec54408c93c2a08285c727af.
