+++ cursor/cursor.go
